first pass at creating uicontroller for LoginActivity
