+++ app/src/org/commcare/dalvik/activities/LoginActivity.java
@@ -4,34 +4,19 @@ import android.annotation.TargetApi;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
-import android.graphics.Bitmap;
-import android.graphics.drawable.ColorDrawable;
-import android.graphics.drawable.StateListDrawable;
 import android.os.Build;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
 import android.support.annotation.NonNull;
 import android.support.v4.app.ActivityCompat;
-import android.text.Editable;
-import android.text.InputType;
-import android.text.TextWatcher;
 import android.util.Log;
 import android.util.Pair;
-import android.util.StateSet;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
-import android.view.View.OnClickListener;
-import android.view.ViewTreeObserver.OnGlobalLayoutListener;
 import android.widget.AdapterView;
 import android.widget.AdapterView.OnItemSelectedListener;
-import android.widget.ArrayAdapter;
-import android.widget.Button;
-import android.widget.CheckBox;
-import android.widget.EditText;
 import android.widget.ImageView;
-import android.widget.Spinner;
-import android.widget.TextView;
 import android.widget.Toast;
 
 import org.commcare.android.database.SqlStorage;
@@ -43,7 +28,6 @@ import org.commcare.android.framework.ManagedUi;
 import org.commcare.android.framework.ManagedUiFramework;
 import org.commcare.android.framework.Permissions;
 import org.commcare.android.framework.RuntimePermissionRequester;
-import org.commcare.android.framework.UiElement;
 import org.commcare.android.javarosa.AndroidLogger;
 import org.commcare.android.models.notifications.MessageTag;
 import org.commcare.android.models.notifications.NotificationMessage;
@@ -59,7 +43,6 @@ import org.commcare.android.tasks.ManageKeyRecordTask;
 import org.commcare.android.tasks.templates.HttpCalloutTask.HttpCalloutOutcomes;
 import org.commcare.android.ui.CustomBanner;
 import org.commcare.android.util.ACRAUtil;
-import org.commcare.android.util.MediaUtil;
 import org.commcare.android.view.ViewUtil;
 import org.commcare.dalvik.R;
 import org.commcare.dalvik.application.CommCareApp;
@@ -94,30 +77,6 @@ public class LoginActivity extends CommCareActivity<LoginActivity>
     private static final int SEAT_APP_ACTIVITY = 0;
     public final static String KEY_APP_TO_SEAT = "app_to_seat";
     public final static String USER_TRIGGERED_LOGOUT = "user-triggered-logout";
-
-    @UiElement(value=R.id.screen_login_bad_password)
-    private TextView errorBox;
-    
-    @UiElement(value=R.id.edit_username, locale="login.username")
-    private EditText username;
-    
-    @UiElement(value=R.id.edit_password, locale="login.password")
-    private EditText password;
-    
-    @UiElement(R.id.screen_login_banner_pane)
-    private View banner;
-
-    @UiElement(value=R.id.login_button, locale="login.button")
-    private Button loginButton;
-
-    @UiElement(value=R.id.restore_session_checkbox)
-    private CheckBox restoreSessionCheckbox;
-
-    @UiElement(R.id.app_selection_spinner)
-    private Spinner spinner;
-
-    @UiElement(R.id.welcome_msg)
-    private TextView welcomeMessage;
     
     private static final int TASK_KEY_EXCHANGE = 1;
     private static final int TASK_UPGRADE_INSTALL = 2;
@@ -128,121 +87,23 @@ public class LoginActivity extends CommCareActivity<LoginActivity>
     private String usernameBeforeRotation;
     private String passwordBeforeRotation;
 
-    private final TextWatcher textWatcher = new TextWatcher() {
-
-        @Override
-        public void beforeTextChanged(CharSequence s, int start, int count, int after) {
-        }
-
-        @Override
-        public void onTextChanged(CharSequence s, int start, int before, int count) {
-        }
-
-        @Override
-        public void afterTextChanged(Editable s) {
-            setStyleDefault();
-        }
-    };
-
-    public void setStyleDefault() {
-        setLoginBoxesColorNormal();
-        username.setCompoundDrawablesWithIntrinsicBounds(getResources().getDrawable(R.drawable.icon_user_neutral50), null, null, null);
-        password.setCompoundDrawablesWithIntrinsicBounds(getResources().getDrawable(R.drawable.icon_lock_neutral50), null, null, null);
-        setupLoginButton();
-        if (loginButton.isEnabled()) {
-            // don't hide error box when showing permission error
-            errorBox.setVisibility(View.GONE);
-        }
-    }
-
-    private void setupLoginButton() {
-        ColorDrawable colorDrawable = new ColorDrawable(getResources().getColor(R.color.cc_brand_color));
-        ColorDrawable disabledColor = new ColorDrawable(getResources().getColor(R.color.grey));
-
-        StateListDrawable sld = new StateListDrawable();
-
-        sld.addState(new int[]{-android.R.attr.state_enabled}, disabledColor);
-        sld.addState(StateSet.WILD_CARD, colorDrawable);
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
-            loginButton.setBackground(sld);
-        } else {
-            loginButton.setBackgroundDrawable(sld);
-        }
-
-        loginButton.setTextColor(getResources().getColor(R.color.cc_neutral_bg));
-    }
+    private LoginActivityUIController uiController;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
-        username.setInputType(InputType.TYPE_TEXT_FLAG_NO_SUGGESTIONS | InputType.TYPE_TEXT_VARIATION_VISIBLE_PASSWORD);
-        setLoginBoxesColorNormal();
+        uiController = new LoginActivityUIController(this);
 
         if (savedInstanceState == null) {
             // Only restore last user on the initial creation
-            SharedPreferences prefs = CommCareApplication._().getCurrentApp().getAppPreferences();
-            String lastUser = prefs.getString(CommCarePreferences.LAST_LOGGED_IN_USER, null);
-            if (lastUser != null) {
-                username.setText(lastUser);
-                password.requestFocus();
-            }
+            uiController.restoreLastUser();
         } else {
             // If the screen was rotated with entered text present, we will want to restore it
             // in onResume (can't do it here b/c will get overriden by logic in refreshForNewApp())
             usernameBeforeRotation = savedInstanceState.getString(KEY_ENTERED_USER);
             passwordBeforeRotation = savedInstanceState.getString(KEY_ENTERED_PW);
         }
-
-        setupUIElements();
-    }
-
-    private void setupUIElements() {
-        username.setInputType(InputType.TYPE_TEXT_FLAG_NO_SUGGESTIONS |
-                InputType.TYPE_TEXT_VARIATION_VISIBLE_PASSWORD);
-
-        setLoginBoxesColorNormal();
-
-        loginButton.setOnClickListener(new OnClickListener() {
-            public void onClick(View arg0) {
-                loginButtonPressed(isRestoreSessionChecked());
-            }
-        });
-
-        username.addTextChangedListener(textWatcher);
-        password.addTextChangedListener(textWatcher);
-
-        username.setHint(Localization.get("login.username"));
-        password.setHint(Localization.get("login.password"));
-
-        final View activityRootView = findViewById(R.id.screen_login_main);
-        final SharedPreferences prefs = CommCareApplication._().getCurrentApp().getAppPreferences();
-        activityRootView.getViewTreeObserver().addOnGlobalLayoutListener(new OnGlobalLayoutListener() {
-            @Override
-            public void onGlobalLayout() {
-                int hideAll = LoginActivity.this.getResources().getInteger(R.integer.login_screen_hide_all_cuttoff);
-                int hideBanner = LoginActivity.this.getResources().getInteger(R.integer.login_screen_hide_banner_cuttoff);
-                int height = activityRootView.getHeight();
-
-                if (height < hideAll) {
-                    banner.setVisibility(View.GONE);
-                } else if (height < hideBanner) {
-                    banner.setVisibility(View.GONE);
-                } else {
-                    // Override default CommCare banner if requested
-                    String customBannerURI = prefs.getString(CommCarePreferences.BRAND_BANNER_LOGIN, "");
-                    if (!"".equals(customBannerURI)) {
-                        Bitmap bitmap = MediaUtil.inflateDisplayImage(LoginActivity.this, customBannerURI);
-                        if (bitmap != null) {
-                            ImageView bannerView = (ImageView) banner.findViewById(R.id.main_top_banner);
-                            bannerView.setImageBitmap(bitmap);
-                        }
-                    }
-                    banner.setVisibility(View.VISIBLE);
-                }
-            }
-        });
 
         Permissions.acquireAllAppPermissions(this, this, Permissions.ALL_PERMISSIONS_REQUEST);
     }
@@ -265,40 +126,35 @@ public class LoginActivity extends CommCareActivity<LoginActivity>
                 for (String requiredPerm : requiredPerms) {
                     if (requiredPerm.equals(permissions[i]) &&
                             grantResults[i] == PackageManager.PERMISSION_DENIED) {
-                        loginButton.setEnabled(false);
-                        errorBox.setVisibility(View.VISIBLE);
-                        errorBox.setText(Localization.get("permission.all.denial.message"));
+                        uiController.setPermissionDeniedState();
                         return;
                     }
                 }
             }
         }
-        loginButton.setEnabled(true);
-        errorBox.setVisibility(View.GONE);
-        errorBox.setText("");
-    }
-
-    private boolean isRestoreSessionChecked() {
-        return restoreSessionCheckbox.isChecked();
+        uiController.setPermissionsGrantedState();
     }
 
     @Override
     public void onSaveInstanceState(Bundle savedInstanceState) {
-        String enteredUsername = username.getText().toString();
+        String enteredUsername = uiController.getEnteredUsername();
         if (!"".equals(enteredUsername) && enteredUsername != null) {
             savedInstanceState.putString(KEY_ENTERED_USER, enteredUsername);
         }
-        String enteredPassword = password.getText().toString();
+        String enteredPassword = uiController.getEnteredPassword();
         if (!"".equals(enteredPassword) && enteredPassword != null) {
             savedInstanceState.putString(KEY_ENTERED_PW, enteredPassword);
         }
     }
 
-    private void loginButtonPressed(boolean restoreSession) {
-        errorBox.setVisibility(View.GONE);
+    /**
+     * @param restoreSession Indicates if CommCare should attempt to restore the saved session
+     *                       upon successful login
+     */
+    protected void initiateLoginAttempt(boolean restoreSession) {
+        uiController.clearErrorMessage();
         ViewUtil.hideVirtualKeyboard(LoginActivity.this);
-
-        DevSessionRestorer.tryAutoLoginPasswordSave(password.getText().toString());
+        DevSessionRestorer.tryAutoLoginPasswordSave(uiController.getEnteredPassword());
 
         if (ResourceInstallUtils.isUpdateReadyToInstall()) {
             // install update, which triggers login upon completion
@@ -323,7 +179,7 @@ public class LoginActivity extends CommCareActivity<LoginActivity>
         // submissions) more centrally.
 
         DataPullTask<LoginActivity> dataPuller = 
-            new DataPullTask<LoginActivity>(getUsername(), password.getText().toString(),
+            new DataPullTask<LoginActivity>(getUniformUsername(), uiController.getEnteredPassword(),
                  prefs.getString("ota-restore-url", LoginActivity.this.getString(R.string.ota_restore_url)),
                  LoginActivity.this) {
                     @Override
@@ -345,7 +201,7 @@ public class LoginActivity extends CommCareActivity<LoginActivity>
                             receiver.raiseLoginMessage(StockMessages.Storage_Full, true);
                             break;
                         case DataPullTask.DOWNLOAD_SUCCESS:
-                            if(!tryLocalLogin(true, isRestoreSessionChecked())) {
+                            if(!tryLocalLogin(true, uiController.isRestoreSessionChecked())) {
                                 receiver.raiseLoginMessage(StockMessages.Auth_CredentialMismatch, true);
                             }
                             break;
@@ -385,7 +241,7 @@ public class LoginActivity extends CommCareActivity<LoginActivity>
                     }
 
                     @Override
-                    protected void deliverError( LoginActivity receiver, Exception e) {
+                    protected void deliverError(LoginActivity receiver, Exception e) {
                         receiver.raiseLoginMessage(StockMessages.Restore_Unknown, true);
                     }
         };
@@ -410,8 +266,8 @@ public class LoginActivity extends CommCareActivity<LoginActivity>
             return;
         }
 
-        // Otherwise, refresh the login screen for current conditions
-        refreshView();
+        // Otherwise, refresh the activity for current conditions
+        refresh();
     }
 
     @Override
@@ -425,24 +281,25 @@ public class LoginActivity extends CommCareActivity<LoginActivity>
         Pair<String, String> userAndPass =
                 DevSessionRestorer.getAutoLoginCreds();
         if (userAndPass != null) {
-            username.setText(userAndPass.first);
-            password.setText(userAndPass.second);
+            uiController.setUsername(userAndPass.first);
+            uiController.setPassword(userAndPass.second);
 
             if (!getIntent().getBooleanExtra(USER_TRIGGERED_LOGOUT, false)) {
                 // If we are attempting auto-login, assume that we want to restore a saved session
-                loginButtonPressed(true);
+                initiateLoginAttempt(true);
             }
         }
     }
 
-    private String getUsername() {
-        return username.getText().toString().toLowerCase().trim();
+    private String getUniformUsername() {
+        return uiController.getEnteredUsername().toLowerCase().trim();
     }
     
     private boolean tryLocalLogin(final boolean warnMultipleAccounts, boolean restoreSession) {
         //TODO: check username/password for emptiness
-        return tryLocalLogin(getUsername(), password.getText().toString(), warnMultipleAccounts,
-                restoreSession);
+        return tryLocalLogin(
+                getUniformUsername(), uiController.getEnteredPassword(),
+                warnMultipleAccounts, restoreSession);
     }
         
     private boolean tryLocalLogin(final String username, String password,
@@ -485,10 +342,9 @@ public class LoginActivity extends CommCareActivity<LoginActivity>
 
                 @Override
                 public void keysLoginComplete(LoginActivity r) {
-                    if(triggerTooManyUsers) {
-                        // We've successfully pulled down new user data.
-                        // Should see if the user already has a sandbox and let
-                        // them know that their old data doesn't transition
+                    if (triggerTooManyUsers) {
+                        // We've successfully pulled down new user data. Should see if the user
+                        // already has a sandbox and let them know that their old data doesn't transition
                         r.raiseMessage(NotificationMessageFactory.message(StockMessages.Auth_RemoteCredentialsChanged), true);
                         Logger.log(AndroidLogger.TYPE_USER, "User " + username + " has logged in for the first time with a new password. They may have unsent data in their other sandbox");
                     }
@@ -607,38 +463,16 @@ public class LoginActivity extends CommCareActivity<LoginActivity>
 
     private void raiseMessage(NotificationMessage message, boolean showTop) {
         String toastText = message.getTitle();
-
         if (showTop) {
             CommCareApplication._().reportNotificationMessage(message);
             toastText = Localization.get("notification.for.details.wrapper",
                     new String[] {toastText});
         }
-        
-        setLoginBoxesColorError();
-        username.setCompoundDrawablesWithIntrinsicBounds(getResources().getDrawable(R.drawable.icon_user_attnneg),  null, null, null);
-        password.setCompoundDrawablesWithIntrinsicBounds(getResources().getDrawable(R.drawable.icon_lock_attnneg), null, null, null);
-        loginButton.setBackgroundColor(getResources().getColor(R.color.cc_attention_negative_bg));
-        loginButton.setTextColor(getResources().getColor(R.color.cc_attention_negative_text));
-
-        errorBox.setVisibility(View.VISIBLE);
-        errorBox.setText(toastText);
-
+        uiController.setErrorMessageUI(toastText);
         Toast.makeText(this, toastText, Toast.LENGTH_LONG).show();
     }
 
-    private void setLoginBoxesColorNormal() {
-        int normalColor = getResources().getColor(R.color.login_edit_text_color);
 
-        username.setTextColor(normalColor);
-        password.setTextColor(normalColor);
-    }
-
-    private void setLoginBoxesColorError() {
-        int errorColor = getResources().getColor(R.color.login_edit_text_color_error);
-
-        username.setTextColor(errorColor);
-        password.setTextColor(errorColor);
-    }
 
     /**
      * Implementation of generateProgressDialog() for DialogController -- other methods
@@ -670,100 +504,50 @@ public class LoginActivity extends CommCareActivity<LoginActivity>
         return dialog;
     }
 
-    private void refreshView() {
+    private void refresh() {
+
         // In case the seated app has changed since last time we were in LoginActivity
-        refreshForNewApp();
+        uiController.refreshForNewApp();
+
+        uiController.updateBanner();
 
         restoreEnteredTextFromRotation();
 
-        updateBanner();
-
         // Decide whether or not to show the app selection spinner based upon # of usable apps
         ArrayList<ApplicationRecord> readyApps = CommCareApplication._().getUsableAppRecords();
+
         if (readyApps.size() == 1) {
-            spinner.setVisibility(View.GONE);
-            welcomeMessage.setText(Localization.get("login.welcome.single"));
             // Set this app as the last selected app, for use in choosing what app to initialize
             // on first startup
             ApplicationRecord r = readyApps.get(0);
             SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(this);
             prefs.edit().putString(KEY_LAST_APP, r.getUniqueId()).commit();
-        }
-        else {
-            welcomeMessage.setText(Localization.get("login.welcome.multiple"));
+
+            uiController.setSingleAppUIState();
+        } else {
             ArrayList<String> appNames = new ArrayList<>();
             appIdDropdownList.clear();
             for (ApplicationRecord r : readyApps) {
                 appNames.add(r.getDisplayName());
                 appIdDropdownList.add(r.getUniqueId());
             }
-            ArrayAdapter<String> adapter = new ArrayAdapter<>(this, R.layout.spinner_text_view, appNames);
-            adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
-            spinner.setAdapter(adapter);
-            spinner.setOnItemSelectedListener(this);
 
-            // Set the spinner's selection to match whatever the currently seated app is
+            // Want to set the spinner's selection to match whatever the currently seated app is
             String currAppId = CommCareApplication._().getCurrentApp().getUniqueId();
             int position = appIdDropdownList.indexOf(currAppId);
-            spinner.setSelection(position);
-            spinner.setVisibility(View.VISIBLE);
-        }
-    }
 
-    private void updateBanner() {
-        ImageView topBannerImageView =
-                (ImageView)banner.findViewById(org.commcare.dalvik.R.id.main_top_banner);
-        if (!CustomBanner.useCustomBannerFitToActivity(this, topBannerImageView)) {
-            topBannerImageView.setImageResource(R.drawable.commcare_logo);
+            uiController.setMultipleAppsUIState(appNames, position);
         }
     }
 
     private void restoreEnteredTextFromRotation() {
         if (usernameBeforeRotation != null) {
-            username.setText(usernameBeforeRotation);
+            uiController.setUsername(usernameBeforeRotation);
             usernameBeforeRotation = null;
         }
         if (passwordBeforeRotation != null) {
-            password.setText(passwordBeforeRotation);
+            uiController.setPassword(passwordBeforeRotation);
             passwordBeforeRotation = null;
-        }
-    }
-
-    private void refreshForNewApp() {
-        // Remove any error content from trying to log into a different app
-        setStyleDefault();
-
-        final SharedPreferences prefs = CommCareApplication._().getCurrentApp().getAppPreferences();
-        String lastUser = prefs.getString(CommCarePreferences.LAST_LOGGED_IN_USER, null);
-        if (lastUser != null) {
-            // If there was a last user for this app, show it
-            username.setText(lastUser);
-            password.requestFocus();
-        } else {
-            // Otherwise, clear the username text so it does not show a username from a different app
-            username.setText("");
-            username.requestFocus();
-        }
-
-        // Clear any password text that was entered for a different app
-        password.setText("");
-
-        // Refresh the breadcrumb bar for new app name
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB) {
-            refreshActionBar();
-        }
-
-        // Refresh UI for potential new language
-        ManagedUiFramework.loadUiElements(this);
-
-        // Refresh welcome msg separately bc cannot set a single locale for its UiElement
-        welcomeMessage.setText(Localization.get("login.welcome.multiple"));
-
-        // Update checkbox visibility 
-        if (DevSessionRestorer.savedSessionPresent()) {
-            restoreSessionCheckbox.setVisibility(View.VISIBLE);
-        } else {
-            restoreSessionCheckbox.setVisibility(View.GONE);
         }
     }
 
@@ -806,7 +590,7 @@ public class LoginActivity extends CommCareActivity<LoginActivity>
                             CommCareApplication._().reportNotificationMessage(NotificationMessageFactory.message(result));
                         }
 
-                        localLoginOrPullAndLogin(isRestoreSessionChecked());
+                        localLoginOrPullAndLogin(uiController.isRestoreSessionChecked());
                     }
 
                     @Override
@@ -823,7 +607,7 @@ public class LoginActivity extends CommCareActivity<LoginActivity>
                                 Localization.get("login.update.install.failure"),
                                 Toast.LENGTH_LONG).show();
 
-                        localLoginOrPullAndLogin(isRestoreSessionChecked());
+                        localLoginOrPullAndLogin(uiController.isRestoreSessionChecked());
                     }
                 };
         task.connect(this);
